Banco de Dados Não-Relacional/N1_ETL_NOSQL/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
pymongo==4.7.2
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
numpy==1.26.4
unicodedata2==15.1.0
psycopg2-binary==2.9.9
